backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
ollama>=0.1.0
chromadb>=0.4.0
pyttsx3>=2.90
pyngrok>=7.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
numpy>=1.24.0
scipy>=1.11.0
python-socketio>=5.10.0
openai>=1.0.0
